Make util.download_url resilient to parallel downloads of the same resource
This was first noticed in #23.
The fix idea is to use a really temporary file instead of a tmp suffixed file for handling conditions of multiple-parallel downloads.

It's out of the scope of the change the addition of logic that would limit the paralel download of the resource. This is justified (at least at the moment) by the fact that the resource download happens only around the time of the bump of the version and for the very first pre-commit run, so it should be acceptable.