Improve bump external releases tool to properly support palantir
palantir-cli releases respect to all the other tools have a name that is more human like and less automation friendly.
Still, releases (also for prior tools) are based on tags that are automation friendly.

In here, we switch over from using the name of the release to use the name of the tag associated to the release